+++ src/jenn/core/model.py
@@ -33,7 +33,7 @@ functions doing computations under-the-hood.
     # Predict response and partials in one step (preferred)
     y_pred, dydx_pred = nn.evaluate(x_test)
 
-.. Note::
+.. note::
     The method `evaluate()` is preferred over separately
     calling `predict()` followed by `predict_partials()`
     whenever both the response and its partials are needed at the same point.
@@ -52,11 +52,11 @@ from typing import Any
 
 import numpy as np
 
-from .core.cache import Cache
-from .core.data import Dataset, denormalize, denormalize_partials, normalize
-from .core.parameters import Parameters
-from .core.propagation import model_forward, model_partials_forward, partials_forward
-from .core.training import train_model
+from .cache import Cache
+from .data import Dataset, denormalize, denormalize_partials, normalize
+from .parameters import Parameters
+from .propagation import model_forward, model_partials_forward, partials_forward
+from .training import train_model
 
 __all__ = ["NeuralNet"]