+++ src/jenn/post_processing/__init__.py
@@ -1,0 +1,23 @@
+"""Post-processing module entry point."""
+# Copyright (C) 2018 Steven H. Berguin
+# This work is licensed under the MIT License.
+
+from ._actual_by_predicted import plot_actual_by_predicted
+from ._contours import plot_contours
+from ._convergence import plot_convergence
+from ._goodness_of_fit import plot_goodness_of_fit
+from ._histogram import plot_histogram
+from ._residual_by_predicted import plot_residual_by_predicted
+from ._sensitivities import plot_sensitivity_profiles
+from .metrics import rsquare
+
+__all__ = [
+    "plot_actual_by_predicted",
+    "plot_contours",
+    "plot_convergence",
+    "plot_goodness_of_fit",
+    "plot_histogram",
+    "plot_residual_by_predicted",
+    "plot_sensitivity_profiles",
+    "rsquare",
+]