+++ src/jenn/utils/__init__.py
@@ -2,13 +2,10 @@
 # Copyright (C) 2018 Steven H. Berguin
 # This work is licensed under the MIT License.
 
-from . import metrics, plot
 from ._jmp import from_jmp
 from ._rbf import rbf
 
 __all__ = [
     "from_jmp",
-    "metrics",
-    "plot",
     "rbf",
 ]